Update _resources.md
We have devoped a very fast screensharing software and dogfood it almost every day for peer programming. AnyDesk went to public beta two weeks ago and we would be happy if the community tried it and gave us feedback :-) Unfortunately, the MacOS version is not finished, yet, so this is still Windows-only. All the best!